Need wrappers around the resume test in scp.sh in case they are building without openssl. This is just to help it pass the CI tests.

+++ regress/scp.sh
@@ -116,33 +116,42 @@ for mode in scp sftp ; do
 	$SCP $scpopts -r somehost:${DIR} ${DIR2} || fail "copy failed"
 	diff ${DIFFOPT} ${DIR} ${DIR2} || fail "corrupted copy"
 
-	verbose "$tag: resume remote dir to local dir"
-	scpclean
-	rm -rf ${DIR2}
-	cp ${DATA} ${DIR}/copy1
-	cp ${DATA} ${DIR}/copy2
-	cp ${DATA} ${DIR}/copy3
-	$SCP -r $scpopts somehost:${DIR} ${DIR2} || fail "copy failed"
-	truncate --size=-512 ${DIR2}/copy1
-	truncate --size=+512 ${DIR2}/copy2
-	$SCP -Z $scpopts somehost:${DIR}/* ${DIR2} || fail "resume failed"
-	for i in $(cd ${DIR} && echo *); do
-            cmp ${DIR}/$i ${DIR2}/$i || fail "corrupted resume copy"
-	done
+	# we don't want to run this test if openssl is not installed
+	# if isn't then the usage text won't have the "vZ" in the
+	# output
+	if $SCP -Z 2>&1 | grep "vZ" > /dev/null 2>&1
+	then
+	    verbose "$tag: resume remote dir to local dir"
+	    scpclean
+	    rm -rf ${DIR2}
+	    cp ${DATA} ${DIR}/copy1
+	    cp ${DATA} ${DIR}/copy2
+	    cp ${DATA} ${DIR}/copy3
+	    $SCP -r $scpopts somehost:${DIR} ${DIR2} || fail "copy failed"
+	    truncate --size=-512 ${DIR2}/copy1
+	    truncate --size=+512 ${DIR2}/copy2
+	    $SCP -Z $scpopts somehost:${DIR}/* ${DIR2} || fail "resume failed"
+	    for i in $(cd ${DIR} && echo *); do
+		cmp ${DIR}/$i ${DIR2}/$i || fail "corrupted resume copy"
+	    done
+	fi
 
-	verbose "$tag: resume local dir to remote dir"
-	scpclean
-	rm -rf ${DIR2}
-	cp ${DATA} ${DIR}/copy1
-	cp ${DATA} ${DIR}/copy2
-	cp ${DATA} ${DIR}/copy3
-	$SCP -r $scpopts ${DIR} somehost:${DIR2} || fail "copy failed"
-	truncate --size=-512 ${DIR2}/copy1
-	truncate --size=+512 ${DIR2}/copy2
-	$SCP -Z $scpopts ${DIR}/* somehost:${DIR2} || fail "resume failed"
-	for i in $(cd ${DIR} && echo *); do
-            cmp ${DIR}/$i ${DIR2}/$i || fail "corrupted resume copy"
-	done
+	if $SCP -Z 2>&1 | grep "vZ" > /dev/null 2>&1
+	then
+	    verbose "$tag: resume local dir to remote dir"
+	    scpclean
+	    rm -rf ${DIR2}
+	    cp ${DATA} ${DIR}/copy1
+	    cp ${DATA} ${DIR}/copy2
+	    cp ${DATA} ${DIR}/copy3
+	    $SCP -r $scpopts ${DIR} somehost:${DIR2} || fail "copy failed"
+	    truncate --size=-512 ${DIR2}/copy1
+	    truncate --size=+512 ${DIR2}/copy2
+	    $SCP -Z $scpopts ${DIR}/* somehost:${DIR2} || fail "resume failed"
+	    for i in $(cd ${DIR} && echo *); do
+		cmp ${DIR}/$i ${DIR2}/$i || fail "corrupted resume copy"
+	    done
+	fi
 
 	verbose "$tag: unmatched glob file local->remote"
 	scpclean
